--- Assets/ChangeSexlogFontSize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c530687d5a0bd964eb240da7d22776bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
