--- Assets/CombatHelp.cs
@@ -1,5 +1,0 @@
-﻿public class CombatHelp : HelpBox
-{
-    private const string saveName = "CombatHelp";
-    protected override string SaveName => saveName;
-}
