--- Assets/CombatHelp.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2996c1a9ace68aa48a67768508f87e61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
