+++ Assets/Editor/PerkInfoEditor.cs
@@ -6,7 +6,7 @@ public class PerkInfoEditor : BaseInfoEditor
 {
     #region Assing SerProps
 
-    private SerializedProperty perk, needCharStat, neededCharStats, needOtherPerks, neededPerks;
+    private SerializedProperty perk, needCharStat, neededCharStats, needOtherPerks, neededPerks, isExcluvsive, exclusiveWith;
 
     private void OnEnable()
     {
@@ -16,12 +16,15 @@ public class PerkInfoEditor : BaseInfoEditor
         neededCharStats = serializedObject.FindProperty("neededCharStats");
         needOtherPerks = serializedObject.FindProperty("needOtherPerks");
         neededPerks = serializedObject.FindProperty("neededPerks");
+        isExcluvsive = serializedObject.FindProperty("isExcluvsive");
+        exclusiveWith = serializedObject.FindProperty("exclusiveWith");
     }
 
     #endregion Assing SerProps
 
     private bool NeedStat { get => needCharStat.boolValue; set => needCharStat.boolValue = value; }
     private bool NeedPerk { get => needOtherPerks.boolValue; set => needOtherPerks.boolValue = value; }
+    private bool Exclusive { get => isExcluvsive.boolValue; set => isExcluvsive.boolValue = value; }
 
     public override void OnInspectorGUI()
     {
@@ -31,6 +34,9 @@ public class PerkInfoEditor : BaseInfoEditor
         EditorGUILayout.BeginHorizontal();
         NeedStat = EditorGUILayout.Toggle("Have stat requirements", NeedStat);
         NeedPerk = EditorGUILayout.Toggle("Have perk requirements", NeedPerk);
+        EditorGUILayout.EndHorizontal();
+        EditorGUILayout.BeginHorizontal();
+        Exclusive = EditorGUILayout.Toggle("Is exclusive perk", Exclusive);
         EditorGUILayout.EndHorizontal();
         if (NeedStat)
         {
@@ -78,6 +84,29 @@ public class PerkInfoEditor : BaseInfoEditor
                 {
                     neededPerks.DeleteArrayElementAtIndex(i);
                 }
+                EditorGUILayout.EndVertical();
+            }
+            EditorGUILayout.EndVertical();
+        }
+        if (Exclusive)
+        {
+            EditorGUILayout.BeginVertical("box");
+            EditorGUILayout.LabelField("If excluvsive with", EditorStyles.boldLabel);
+            if (GUILayout.Button("Add"))
+            {
+                exclusiveWith.arraySize++;
+            }
+            for (int i = 0; i < exclusiveWith.arraySize; i++)
+            {
+                EditorGUILayout.BeginVertical("box");
+                SerializedProperty ele = exclusiveWith.GetArrayElementAtIndex(i);
+                EditorGUILayout.BeginHorizontal();
+                EditorGUILayout.PropertyField(ele,GUIContent.none);
+                if (GUILayout.Button("Remove"))
+                {
+                    exclusiveWith.DeleteArrayElementAtIndex(i);
+                }
+                EditorGUILayout.EndHorizontal();
                 EditorGUILayout.EndVertical();
             }
             EditorGUILayout.EndVertical();
