--- Assets/GameUIHelpBox.cs
@@ -1,6 +1,0 @@
-﻿public class GameUIHelpBox : HelpBox
-{
-    private const string saveName = "GameUIHelp";
-
-    protected override string SaveName => saveName;
-}
