--- Assets/GameUIHelpBox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63cd40ecb851a6544843e8bb6615a7f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
