--- Assets/GenetricChangeFontSize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a8b28ac926a0d0d44833c0269eddd836
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
