--- Assets/HelpBox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60bf7786f75642243a3e5503699c4816
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
