+++ Assets/Scripts/ChangeCombatFontSize.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45372c655b714c24688c8baafca44f95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
