+++ Assets/Scripts/ChangeFontSize.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68c6793f862f65d4bb2f08f0344fec65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
