+++ Assets/Scripts/ChangeSexlogFontSize.cs
@@ -1,0 +1,8 @@
+﻿public class ChangeSexlogFontSize : ChangeFontSize
+{
+    protected override float CurrSize => Settings.SexlogFontSize;
+
+    protected override float DownSized => Settings.SexlogFontSizeDown;
+
+    protected override float UpSized => Settings.SexlogFontSizeUp;
+}
