+++ Assets/Scripts/CombatHelp.cs
@@ -1,0 +1,5 @@
+﻿public class CombatHelp : HelpBox
+{
+    private const string saveName = "CombatHelp";
+    protected override string SaveName => saveName;
+}
