+++ Assets/Scripts/GameUIHelpBox.cs
@@ -1,0 +1,6 @@
+﻿public class GameUIHelpBox : HelpBox
+{
+    private const string saveName = "GameUIHelp";
+
+    protected override string SaveName => saveName;
+}
