+++ Assets/Scripts/Menus/PerkTree/Perk infos/Vore.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eaa3899b779817445ba747a770c398b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
